requirements.txt
requests==2.11.1
PixivPy==3.1.0
